synced adding friends by moving logic to parent
